refactor(teleop): use existing blueprints for Quest HUD
